billing/collector: rename service name.
    - rename exporter service to collector in order to avoid service
    name conflict.
