+++ systems/billing/collector/Dockerfile
@@ -1,0 +1,5 @@
+FROM alpine:3.13.5
+
+COPY bin/collector /usr/bin/collector
+
+CMD ["/usr/bin/collector"]
