+++ systems/billing/collector/Int.Dockerfile
@@ -1,4 +1,4 @@
 FROM 003664043471.dkr.ecr.us-east-1.amazonaws.com/e2e-base-image:latest
 
-ENV PROJECT_NAME=exporter
+ENV PROJECT_NAME=collector
 COPY bin/integration /usr/bin/integration
