+++ systems/billing/collector/cmd/version/version.go
@@ -1,0 +1,5 @@
+package version
+
+// Minor version is autoupdated by the build system
+// NOTE: use go build -ldflags "-X github.com/ukama/ukama/systems/billing/collector/cmd/version.Version==$(git describe)".
+var Version = "v0.0.debug"
