--- systems/billing/exporter/Dockerfile
@@ -1,5 +1,0 @@
-FROM alpine:3.13.5
-
-COPY bin/exporter /usr/bin/exporter
-
-CMD ["/usr/bin/exporter"]
