--- systems/billing/exporter/cmd/version/version.go
@@ -1,5 +1,0 @@
-package version
-
-// Minor version is autoupdated by the build system
-// NOTE: use go build -ldflags "-X github.com/ukama/ukama/systems/billing/exporter/cmd/version.Version==$(git describe)".
-var Version = "v0.0.debug"
